Document matrix multiplication operators (#6852)
* Document matrix multiplication operators

This is to address one of cases where a user was confused about why
Slang doesn't have operator overloading of multiplication for matrices.

It is little confusing that glsl.meta.slang implemented `operator*()`
for matrices but Slang/HLSL doesn't. It turned out that HLSL/Slang
performs a component-wise multiplication with `operator*()` whereas GLSL
performs a matrix multiplication.

* Mentioning CUDA

* Update based on the review feedback