Cleanups related to RIFF support
This change ends up being an almost complete rewrite of the subsystem in `slang-riff.{h,cpp}`.

The original design of `RiffContainer` had a single type servicing the cases when the code wants to *read* a RIFF as well as when it wants to *write* a RIFF.
It also contained a lot of API surface area, much of which only got used in a small number of places, if at all.

The biggest change in the new implementation is that there is a strong separation between reading and writing of RIFF-structured data.

Reading
-------

When reading, the new `RIFF::Chunk` type can be used to traverse the hierarchy of a RIFF-structured file in memory without needing to make a copy of it.
One concrete consequence of this change is that we no longer make extra copies of the entire serialized core module (one copy of the compressed version, and then one copy of the decompressed version), as part of creating a global session;
instead, the code is able to directly traverse the RIFF hierarchy in the `ISlangBlob` that the archive file system returns.

The implementation includes a fair bit of validation logic to try to make sure that code is defensive against accidentally corrupted or malformed input (for example, see `RIFF::RootChunk::getFromBlob()`, `RIFF::ListChunk::getFirstChild()`, and `RIFF::BoundsCheckedChunkPtr::getNextSibling()`), although it is certainly not hardened against malicious input.

Because reading of RIFF data is now being done without any copying, any code reading from RIFF chunks needs to be robust against unaligned data.
Fortunately, there were very few locations in the codebase that were assuming alignment when reading from RIFF files, and those few locations were easy enough to fix.

Some of the code related top deserialization of Slang modules was already using wrapper types around pointers to chunks in a `RiffContainer` (e.g., `ModuleChunkRef`).
These use cases could be conveniently ported to declare subtypes of `RIFF::Chunk`, which allows all of that code to pass around pointer directly into the serialized RIFF data, while also maintaining some static typing so that we can clearly tell, e.g., an `ASTModuleChunk` from an `IRModuleChunk`.

Writing
-------

What used to be the `RiffContainer` type is now split between `RIFF::Builder`, which owns the tree of chunks, and `RIFF::BuildCursor`, which factors out the stateful part of `RiffContainer` where it had a notion of a "current chunk."

The builder uses a hierarchy of `RIFF::ChunkBuilder` types, distinct from the `RIFF::Chunk` types that are used for reading.

A lot of the functionality has been stripped out from this version.
The hope is that what remains is easier to understand, and will be easier to maintain going forward.

Other Stuff
-----------

Smaller changes include:

* Made the RIFF `FourCC` type use a `union` so that it is easier to see the textual version of a `FourCC` when debugging.

* Took a lot of the RIFF-related functionality that was only being used by `unit-test-riff.cpp` and moved it into the unit test file itself. If we ever want that stuff to be public API again, we could migrate the code back, but I don't see that happening.

* Removed the `RiffSemanticVersion` type, since it was only being used by `slang-repro.{h,cpp}`, which could easily just use `SemanticVersion`. Also cleaned up some other use cases or `SemanticVersion` based on changes to its exposed API.

* Added a `Size` type to `slang-common.h` to go along with `Index` and `Count` in terms of documenting intent. Dealt with the fallout around `slang-gfx` and `slang-rhi` already defining their own `Size` types.

* Changed the `Byte` type in `slang-common.h` to be `unsigned char`, to help us try to avoid undefined behavior when doing type punning using arrays of `Byte`s.