Support Array Sizes using Generic arguments to be initialized via {} (#6720)
* Add support for Array Sizes using Generic arguments to be initialized via {}

Fixes one subissue of #6138

This change adds support for initializing Arrays with Generic size arguments via {}
and adds a test to verify it.

The change checks for an array whose size parameter is a GenericParamIntVal
and since the size of such an array will be known at link time, is not considered
as a case of the size not being known statically.

* Add support for Array Sizes using Generic arguments to be initialized via {}

Fixes one subissue of #6138.
Fixes the issue #6958.

This change adds support for initializing Arrays with Generic size arguments via {}
and adds a test to verify it.

Support is added by means of adding a new AST Expr node that lowers down to the IR MakeArrayFromElement
and the emission of a diagnostic is replaced with the creation of this new AST Expr node.

* format code

---------

Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>