Fix slangc path used by cmdline reference doc regeneration action (#7128)
This change uses a different path for accessing the slangc binary during the command line reference doc regeneration action, as the environment variable referenced for the bin directory may not be accurate.

This change also makes some exclusions from the slang build in the command line reference actions, to speed up those builds for checking/regenerating.