Add command line option for separate debug info
Add command line arg -separate-debug-info which, if provided, produces
both a .spv and a .dbg.spv file. The .dbg.spv file contains full debug
info and the .spv file has all debug info stripped out.

This patch also updates external/spirv-tools to point at the slang
fork which has a patch needed for this to work. We will change this
back when the patch is upstreamed.