Fix missing debug info in the included slang file
Issue:
https://github.com/shader-slang/slang/issues/7271

Debug info including DebugFunction, DebugLocation, and DebugValue
are missing in IR for "#included" Slang shader file.

The included shader file was not added to TranslationUnit's source
file list, therefore mapSourceFileToDebugSourceInst.add() was not
called for the source in generateIRForTranslationUnit(), and later
mapSourceFileToDebugSourceInst.tryGetValue() could not get value
for the source to add DebugLocationDecoration, which led to missing
DebugFunction, DebugLocation and other debug info for the included
file in IR.

Adding the include file in TranslationUnit's source file list fixes
the issue.