Implement isnan and isinf for WGSL with bitwise operations (#7344)
WGSL doesn't support isnan and isinf, because it assumes that it always uses fast-math and fast-math doesnt' handle NaN as defined in IEEE standard.

The initial implementation used a clever workaround but it stopped working from some point.

This PR implemented isnan and isinf with a bitwise operation, which can be expensive.
But that seems to be an only option at the moment.