AD cleanup for 'break' replacement (#7456)
close #4187.

This PR makes some cleanup on the variable name.
During CFG normalization, we introduce a variable to replace the break, but the variable indicates that we should continue running the loop if it's true, and break the loop if it's false.

Previously we named this variable as 'breakVar', which is opposite to its purpose. So rename it to keepRunning.