Fix the invalid SPIRV decoration issue
Close #7508.

SPIRV doesn't allow decoration on type with Private or Function storage
class.

In our lowering logic, if the array type is used by buffer type it
will always have stride operand after lowering, so if the array is not
used by buffer type, it must be used only by thread_local or
group_shared variable, and it will not have stride operand. For this
case, we don't need to emit stride decoration for SPIRV.