Correct the scope for DebugLocalVariable (#7516)
* Correct the scope for DebugLocalVariable

All of the DebugLocalVariable instructions have their scope
set to DebugCompilationUnit. The scope should instead be set
to a DebugFunction, or a DebugLexicalBlock which is recursively
inside a DebugFunction.

Register the debug info for the function instructions, which
helps findDebugScope() to find the right DebugFunction scope
for DebugLocalVariable.

* Add a test for DebugLocalVariable

To check the scope, which should be a DebugFunction.